first cut at building on osx
Went pretty smooth, auto* added to README.

Only one main issue:

Character storage issue surfaced in `test_02`.
Not sure what the deal is with this, because it's late
and I'm tired. So it's #if 0'ed for now.

A very slight issue is on a printf format specifier,
warning on compile, simple: s/d/zy/ fix. Hopefully this
works on all platforms. If not casting to (int) or
just ignoring the warning works just as well.

Other than that, everything looks pretty good.

If I can figure out what the problem is for `test_02`
and can fix it, I'll issue a pull request.